PassRolestoMediaConvert Sid was
Using this was giving me the error `This policy contains the following error: The following resources are invalid. They must be either '*' or an arn pattern: arn:aws:iam::*`,  for the `PassRolestoMediaConvert Sid`. Your version does have the `arn:aws:iam::*` pattern, but I couldn't save the role until I changed this to `*`